Bump SolrJ from 8.11.2 to 9.3.0
With CMS 2307.1, SolrJ has been updated from 8.11.2 to 9.2.1. As the
Spring Boot Parent still manages the Solr versions to 8.11.2, we must
update the management for this project to Solr 9.2.1, using the same
property `solr.version` in order to prevent dependency conflicts. It
should be safe to even go with the latest minor version. And we can
now rely on Dependabot updates instead of having to manually update
SolrJ after each and every CMS update.